requirements.txt
# 주식자동매매프로그램 의존성 패키지
pandas>=2.2.0
numpy>=1.26.0
requests>=2.31.0
websocket-client>=1.6.4
python-dotenv>=1.0.0
schedule>=1.2.0
matplotlib>=3.8.0
seaborn>=0.13.0
ta>=0.10.2
yfinance>=0.2.28
ccxt>=4.1.77
sqlalchemy>=2.0.23

# 웹 대시보드 관련 패키지
fastapi>=0.104.0
uvicorn>=0.24.0
jinja2>=3.1.0
python-multipart>=0.0.6
aiofiles>=23.0.0 